requirements.txt
openai==1.14.2
Pillow==10.2.0
python-dotenv==1.0.1
streamlit==1.32.0
